requirements.txt
ragas>=0.1.0
openai>=1.0.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain>=0.1.0
nest-asyncio>=1.5.0
